play around with implicit casts now

+++ trufflesqueak/de.hpi.swa.trufflesqueak/src/de/hpi/swa/trufflesqueak/nodes/primitives/impl/PrimBitAnd.java
@@ -11,7 +11,7 @@ public class PrimBitAnd extends PrimitiveBinaryOperation {
     }
 
     @Specialization
-    protected int bitAnd(int receiver, int arg) {
+    protected long bitAnd(long receiver, long arg) {
         return receiver & arg;
     }
 }
